Added support for new lines for link lists split
The fields for link lists (documentation and sample links) are text areas and not single line fields. While there is a note below the field that directs the editor to use commas to separate the links, this is confusing and nonsensical in text areas or links (there will never be new lines within URLs) and makes it error prone, as evident in https://groups.google.com/a/chromium.org/forum/#!topic/blink-dev/BdAbzHMdjh8 - and so support for new lines as well as commas was added (I prefer to just use new lines, though).