Fix metrics view to align columns across all browsers (#322)
Apply flex via simple positional rules that apply to all rows of the table
instead of trying to name the elements in both the header and body that
should get flex: 1 vs. flex: 2.  The precentage heading was matching
both rules and the specificity differed depending on whether we had
native ShadowDOM.  Fixes #301