Distinguished Chrome for iOS from other versions
Some entries creep in that mention that a feature is supported on Chrome for iOS when it does not. Instead of removing it completely, since a few features are apparently polyfilled by Chrome for iOS, the fields were re-ordered a bit to have it as the last Chrome-related field (right behind the Opera section) and the help text was prepended with some wording to make it more noticable that it is Chrome for iOS and that it usually does not include those features.