Inintial introduction of Blink Launch Process data into Chromestatus.
Add a set of fields to the entry creation/edit form so that we collect
all the information that is part of the blink launch process. This
also makes chromestatus generate up to date intent templates with the
future intent of eliminating the need to manually create and send
intent emails. This includes making feature entries aware of the
current intent stage they're in as well as holding on to all the
additional data so that creating intent messages for later stages of
the launch process is a matter of updating the entry and generating a
new intent message.