Expand the length of the free form text fields
This expands the length of the free form text fields from 500 to 1480.
There's a hard maximum of 1500 in the database, and it seems the front
end text field limiting code is buggy and doesn't correctly account for
newlines in the form fields, so I cut this a bit short of the hard limit
to help users more often know that they've reached the limit. This
change also loosens the restriction for the intent to implement link as
users don't always have one at this point.