Make intent template point to a standalone feature page.
Right now the intent template points to feature entries as part of the
list of all entries. Given the non-trivial length of the list of
features today, it seems it makes more sense to make the intent template
point directly to the feature itself instead of just expanding the feature
in the long list of all features.