Implement starred feature models and JSON API handler. (#738)
ChromeStatus currently has a push notifications that pop up a small window whenever someone edits a feature that the user has clicked the bell icon for. We want to phase that out and instead offer star icons that trigger notification emails. See design doc at go/chromestatus-notification-emails.

In this CL:
* Implement feature star datastore entity model classes with methods to star, unstar, and list.
* Implement JSON request handlers to star, unstar, and list starred features.
* Some tweaks to unit test set-up because these are the first significant user tests.