Add rendered templates for unittests
Also, move the TESTDATA functionality to a testing_config to be a test
helper. In this new helper, users can still retreive from it like a
dictionary. In addition, there's a helper function to help make goldens
that can be used to regenerate the golden templates. Add commented out
code to use it.