Show useful error on feature not found
Also, stop showing skeleton upon feature not being found.

Add two new Errors to cs-client.js:
- ChromeStatusHttpError
- FeatureNotFoundError

ChromeStatusHttpError is meant to be returned at the doFetch level
FeatureNotFoundError is meant at the getFeature level. getFeature
interprets ChromeStatusHttpError and returns as necessary.

Add isFeatureLoaded to chromedash-feature-page.
Allows us to detect the case that we have finished loading but nothing was loaded.

Use that function to detect when to show main feature components or show nothing.

In the event a feature is not found. A blank page with the normal header is shown. A toast appears and says:
`Feature not found.`

TODO: add a unit test for chromedash-feature-page.

Fixes #2257