Show useful error on feature not found (#2503)
* Show useful error on feature not found

Also, stop showing skeleton upon feature not being found.

Add two new Errors to cs-client.js:
- ChromeStatusHttpError
- FeatureNotFoundError

ChromeStatusHttpError is meant to be returned at the doFetch level
FeatureNotFoundError is meant at the getFeature level. getFeature
interprets ChromeStatusHttpError and returns as necessary.

Add isFeatureLoaded to chromedash-feature-page.
Allows us to detect the case that we have finished loading but nothing was loaded.

Use that function to detect when to show main feature components or show nothing.

In the event a feature is not found. A blank page with the normal header is shown. A toast appears and says:
`Feature not found.`

TODO: add a unit test for chromedash-feature-page.

Fixes #2257

* clean up conditionals

* Add reset function to toast to clean toast state.

* Add test for No Feature Found

Also, manage toastEl as a singleton for the tests.
Without doing this, the test will get a new element but the browser will use the first element and assertions will fail

* error comparison works now

* Add static text for feature not found error.

Also, revert any unnecessary helpers