devcontainer: run `npm run setup` again
Recently, when running `npm run clean-setup` in the container,
it has not been installing the packages in the virtual environment.
Instead, it has been installing in the default Python environment. We
will need to call npm run setup again once the environment has been
activated.

This problem hasn't been reported for local development setups.
This is a workaround to not mess up the local development setup.

By running npm run setup again, the container doesn't need to download
the packages again. The packages are cached and are just copied over. So
it doesn't increase setup time that much.