Remove context & failed attempt at global
changes:
- Remove the context
- Move the openapi client to js-src
- modify gulp task to:
  - ignore the test file
  - add the openapi-client to the rollup task instead of the js task since it is a es6 module (because it has an import / require statement)
- Attempt to construct the object in the spa template. It failed to work at load time. More details below:

When trying to call my exported class at page load like the regular client, I get a not defined error.
There are thoughts about this:
- https://stackoverflow.com/a/75026222
- https://stackoverflow.com/a/53630402
- https://stackoverflow.com/a/69888825

This commit serves as an attempt to try it out

Turns out, script type "module" is automatically set to defer and cannot be disabled.