Use default empty string for USER_EMAIL in UNIT_TEST_MODE (#2966)
* Sign in user for metricsdata test

Bug:
If a developer runs:

```
python3.10 -m unittest api/metricsdata_test.py
```

It will fail. The error is:

```
Traceback (most recent call last):
  File "/workspaces/chromium-dashboard/api/metricsdata_test.py", line 153, in test_get_template_data
    actual_datapoints = self.handler.get_template_data()
  File "/workspaces/chromium-dashboard/api/metricsdata.py", line 216, in get_template_data
    return super(CSSPopularityHandler, self).get_template_data()
  File "/workspaces/chromium-dashboard/api/metricsdata.py", line 184, in get_template_data
    return _datapoints_to_json_dicts(properties)
  File "/workspaces/chromium-dashboard/api/metricsdata.py", line 37, in _datapoints_to_json_dicts
    user = users.get_current_user()
  File "/workspaces/chromium-dashboard/framework/users.py", line 193, in get_current_user
    if os.environ['USER_EMAIL']!= '':
  File "/usr/local/lib/python3.10/os.py", line 680, in __getitem__
    raise KeyError(key) from None
KeyError: 'USER_EMAIL'
```

Troubleshooting:

The test passes if you run npm run test / npm run do-tests though.
Added a print statement and it turns out when running npm run test, os.environ['USER_EMAIL'] is equal to ''. That is why it is not caught in the regular CI test.

That means there's some test somewhere that runs beforehand that resets os.environ['USER_EMAIL'] to '' during the full suite (most likely the sign_out from testing_config.py which is used in tests).

This change:

Add a sign in on setup and signout on teardown. This will solve when a developer runs a one-off of that test file. It will work.

An alternative that could work:

At the beginning of testing_config.py, set os.environ['USER_EMAIL'] to ''. Open to that too.

* use jrobbins suggestion