Preserve casing in the OpenAPI JS bindings.
It seems like it'll be confusing to have the API documentation use a
different casing from the code we actually write. If we want a
particular casing, we should use it in the API spec.

I left the paramNaming as camelCase because it doesn't seem to react to
changes in the API casing (a bug in the generator?), and we don't want
"ComponentUsersRequest" to start with a capital when used as an object
member.