Playwright db run on port different from 15606 (#3642)
* Playwright db run on port different from 15606

Background:

When the playwright tests were created, it was only tested for users that
did not use the devcontainer. The devcontainer starts a database up for
users automatically which occupies port 15606. The playwright tests spin
up a clean database. But previously fails in devcontainers because the
port 15606 is occupied.

This change:

Modifies the base database used in both devcontainer and playwright docker compose files
to allow the DB_PORT to be configured. The devcontainer will continue to
spin up a database at 15606. The playwright tests will create a database
on port 15607. This prevents the conflict on devcontainer.

This also benefits non-devcontainer users because previously they had to
manually shut down their local emulators so that the playwright database
could run.

* use DATASTORE_EMULATOR_HOST if provided, else use default