Configure formatting with Prettier via gts.
This change disables formatting for a few files with too many changes,
for some tests, and for a few HTML elements where a
`white-space:pre-wrap` CSS rule made whitespace significant.

It also updates the eslint rules to use gts except for rules the
codebase doesn't follow yet, and gts's use of eslint-plugin-prettier.