Make the sidebar fill the vertical space. (#3864)
Also remove a margin-top from the footer that was creating a gap with the sidebar in short windows. Then the footer's shadow disappeared because the z-index was being ignored, which required adding a position property. It's possible we'll want the content-flex-wrapper to include some padding so things don't run right up to the footer, but nothing I've tested looks particularly bad.

This also expands a section in chromedash-guide-feature-page_pwtest.js so the masked history section isn't under the dialog that's being screenshotted.